chore: renamed folder TokenBalances => DashboardTokenBalances for consistency
