+++ apps/extension/src/ui/domains/Portfolio/AssetDetails/DashboardTokenBalances/ChainTokenBalancesUniswapV2Row.tsx
