+++ apps/extension/src/ui/domains/Portfolio/AssetDetails/DashboardTokenBalances/LockedExtra.tsx
