+++ apps/extension/src/ui/domains/Portfolio/AssetDetails/DashboardTokenBalances/TokenBalancesDetailRow.tsx
